assert view degrade to const tests post scheduler graph_rewrite [pr]
